Move move processing functions to moveService utility module and update imports accordingly
